+++ scvi/module/base/_decorators.py
@@ -2,6 +2,7 @@ from collections.abc import Mapping, Sequence
 from functools import wraps
 from typing import Any, Callable, Union
 
+import flax.linen as nn
 import torch
 from torch.nn import Module
 
@@ -118,3 +119,18 @@ def _apply_to_collection(
 
     # data is neither of dtype, nor a collection
     return data
+
+
+def flax_configure(cls: nn.Module) -> Callable:
+    """Decorator to raise an error if the model is in latent mode."""
+    original_init = cls.__init__
+
+    @wraps(original_init)
+    def init(self, *args, **kwargs):
+        self.configure()
+        original_init(self, *args, **kwargs)
+        if not isinstance(self.training, bool):
+            raise ValueError("Custom sublclasses must have a training parameter.")
+
+    cls.__init__ = init
+    return cls
