+++ scvi/utils/_decorators.py
@@ -1,8 +1,6 @@
 from functools import wraps
 from typing import Callable
 
-import flax.linen as nn
-
 
 def unsupported_in_latent_mode(fn: Callable) -> Callable:
     """Decorator to raise an error if the model is in latent mode."""
@@ -16,18 +14,3 @@ def unsupported_in_latent_mode(fn: Callable) -> Callable:
         return fn(self, *args, **kwargs)
 
     return wrapper
-
-
-def flax_configure(cls: nn.Module) -> Callable:
-    """Decorator to raise an error if the model is in latent mode."""
-    original_init = cls.__init__
-
-    @wraps(original_init)
-    def init(self, *args, **kwargs):
-        self.configure()
-        original_init(self, *args, **kwargs)
-        if not isinstance(self.training, bool):
-            raise ValueError("Custom sublclasses must have a training parameter.")
-
-    cls.__init__ = init
-    return cls
